+++ resources/lib/tests/test_addon.py
@@ -1,10 +1,64 @@
 # coding=utf-8
 
 import addon  # test target
-from _testdata import *  # test data
-
-import pytest
 import mock
+import pytest
+
+try:
+    # sensitive test data
+    from _testdata import *
+except ImportError:
+    TESTDATA_VKUSER_LOGIN = ''
+    TESTDATA_VKUSER_PSWD = ''
+    TESTDATA_FP_PROFILE = ''
+    TESTDATA_SEARCHVIDEOS_QUERY = ''
+    TESTDATA_VIDEO = {}
+    TESTDATA_COMMUNITY = {}
+    TESTDATA_LIKEDCOMMUNITY = {}
+
+
+# mocks -----
+
+
+MOCK_ADDONINFO = {
+    'id': 'plugin.video.vk',
+    'name': 'VK',
+}
+
+MOCK_KODIENV = {
+    'fp_profile': TESTDATA_FP_PROFILE,
+    'sysargv': {
+        'path': str(__file__),
+        'handle': int(0),
+        'qs': str(''),
+    }
+}
+
+MOCK_USERSETTINGS = {
+    'forcevideoviewmode': 'true',
+    'forcevideoviewmodeid': '500',
+    'itemsperpage': '200',
+    'keepaddonrequesthistory': 'false',
+    'keepplayedvideohistory': 'false',
+    'preferhls': 'false',
+    'searchadult': 'true',
+    'searchduration': '0',
+    'searchdurationmins': '',
+    'searchown': 'false',
+    'searchsort': '2',  # '0'=bydate, '1'=byduration, '2'=byrelevance
+    'vkuserlogin': TESTDATA_VKUSER_LOGIN,
+    'vkuseraccesstoken': '',
+}
+
+MOCK_USERINPUTS = {
+    'LSTR_30030': TESTDATA_VKUSER_LOGIN,
+    'LSTR_30031': TESTDATA_VKUSER_PSWD,
+    'LSTR_30083': TESTDATA_SEARCHVIDEOS_QUERY,
+}
+
+MOCK_VIDEO = TESTDATA_VIDEO
+MOCK_COMMUNITY = TESTDATA_COMMUNITY
+MOCK_LIKEDCOMMUNITY = TESTDATA_LIKEDCOMMUNITY
 
 
 # patches -----
@@ -22,13 +76,13 @@ def patch_xbmc(monkeypatch, context=addon):
 def patch_xbmcaddon(monkeypatch, context=addon):
 
     def mock_getaddoninfo(id_):  # type: (str) -> str
-        return TEST_ADDONINFO[id_]
+        return MOCK_ADDONINFO[id_]
 
     def mock_getsetting(id_):  # type: (str) -> str
-        return TEST_SETTINGS[id_]
+        return MOCK_USERSETTINGS[id_]
 
     def mock_setsetting(id_, value):  # type: (str, str) -> None
-        TEST_SETTINGS[id_] = value
+        MOCK_USERSETTINGS[id_] = value
 
     def mock_getlocalizedstring(id_):  # type: (int) -> str
         return 'LSTR_{}'.format(id_)
@@ -56,7 +110,7 @@ def patch_xbmcplugin(monkeypatch, context=addon):
 def patch_xbmcgui(monkeypatch, context=addon):
 
     def mock_input(heading, **_):  # type: (str, dict) -> str
-        return TEST_INPUTS[heading]
+        return MOCK_USERINPUTS[heading]
 
     mock_dialogclass = mock.Mock(name='Dialog')
     mock_dialogclass.input.side_effect = mock_input
@@ -72,7 +126,7 @@ def patch_buildfp(monkeypatch, context=addon):
 
     def mock_buildfp(filename):  # type: (str) -> str
         import os
-        fp = str(os.path.join(TEST_ENV['fp_profile'], filename))
+        fp = str(os.path.join(MOCK_KODIENV['fp_profile'], filename))
         return fp
 
     monkeypatch.setattr(context, 'buildfp', mock_buildfp)
@@ -83,7 +137,7 @@ def patch_buildfp(monkeypatch, context=addon):
 def patch_parsesysargv(monkeypatch, context=addon):
 
     def mock_parsesysargv():  # type: () -> dict
-        return TEST_ENV['sysargv']
+        return MOCK_KODIENV['sysargv']
 
     monkeypatch.setattr(context, 'parsesysargv', mock_parsesysargv)
 
@@ -93,8 +147,8 @@ def patch_parsesysargv(monkeypatch, context=addon):
 
 def test_addonglobal(context=addon):
     context.ADDON = context.initaddon()
-    assert context.ADDON.getAddonInfo('id') == TEST_ADDONINFO['id']
-    assert context.ADDON.getSetting('itemsperpage') == TEST_SETTINGS['itemsperpage']
+    assert context.ADDON.getAddonInfo('id') == MOCK_ADDONINFO['id']
+    assert context.ADDON.getSetting('itemsperpage') == MOCK_USERSETTINGS['itemsperpage']
 
 
 def test_initvkauthsession(context=addon):
@@ -121,7 +175,7 @@ def test_vkapi_video_get(context=addon):
     vkapi = context.initvkapi()
     r = vkapi.video.get(
         album_id=None,
-        videos=TEST_VIDEO['_oidid'],
+        videos=MOCK_VIDEO['_oidid'],
         extended=True,
     )
     assert r['count'] == 1
@@ -156,14 +210,14 @@ def test_searchvideos(context=addon):
 
 def test_listsearchedvideos(context=addon):
     context.ADDON = context.initaddon()
-    context.listsearchedvideos(q='test', offset=0)
+    context.listsearchedvideos(q=TESTDATA_SEARCHVIDEOS_QUERY, offset=0)
     assert True
 
 
 def test_listvideos(context=addon):
     context.ADDON = context.initaddon()
     context.listvideos(ownerid=0, albumid=0, offset=0)
-    context.listvideos(ownerid=-TEST_COMMUNITY['id'], albumid=0, offset=0)
+    context.listvideos(ownerid=-MOCK_COMMUNITY['id'], albumid=0, offset=0)
     assert True
 
 
@@ -181,7 +235,7 @@ def test_listwatchlist(context=addon):
 
 def test_playvideo(context=addon):
     context.ADDON = context.initaddon()
-    context.playvideo(ownerid=TEST_VIDEO['owner_id'], videoid=TEST_VIDEO['id'])
+    context.playvideo(ownerid=MOCK_VIDEO['owner_id'], videoid=MOCK_VIDEO['id'])
     assert True